Get n-grams

Query to get n-grams of specified size (n) and type. Each row has:

Speaker's role.

The n-gram (word, stem, or part-of-speech). See CHAT manual for part-of-speech code values.

Frequency count of n-gram.

getNgrams(
  nGram = NULL,
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL,
  auth = FALSE
)

Arguments

nGram
Query for n-grams by size (n) and type. For example, to search for all n-grams of length 3 of word type: nGram=c(size="3", type="word"). Legal value for size is any positive integer equal to or greater than 1. Legal value for type is "word" to return exact word n-grams, "stem" to return word stem n-grams, "pos" to return part of speech n-grams.

corpusName
Name of corpus to query. For example, to search within the childes corpus, corpus="childes".

corpora
Name of corpus/corporas to query. This is a path starting with the corpus name followed by subfolder names leading to a folder for which all transcripts beneath it will be queried. For example, to query all transcripts in the MacWhinney childes corpus: corpora = c('childes', 'Eng-NA', 'MacWhinney').

lang
Query by language For example, to get transcripts that contain both English and Spanish: lang=c("eng", "spa"). Legal values: 3-letter language codes based on the ISO 639-3 standard.

media
Query by media type. For example, to get transcripts with an associated video recording: media=c("video"). Legal values: "audio" or "video".

age
Query by participant month age range. For example, to get transcripts with target participants who are 14-18 months old: age=c(from="3", to="12").

gender
Query by participant gender. For example, to get transcripts with female target participants: gender=c("female"). Legal values: "female" or "male".

designType
Query by design type. For example, to get transcripts from a longitudinal study: designType=c("long") Legal values are "long" for longitudinal studies, "cross" for cross-sectional studies.

activityType
Query by activity type. For example, to get transcripts where the target participant is engaged in toy play: activityType=c("toyplay"). See the CHAT manual for legal values.

groupType
Query by group type. For example, to get transcripts where the target participant is hearing limited: groupType=c("HL"). See the CHAT manual for legal values.

auth
Determine if user should be prompted to authenticate in order to access protected collections. Defaults to False.

Examples

getNgrams(nGram=c("3", "word"),
          corpusName = 'childes',
          corpora = c('childes',
                              'Eng-NA',
                              'MacWhinney',
                              '010411a'));

#> [1] "Fetching data, please wait..."
#> [1] "Success!"

#>             role                        ngram count
#> 1   Target_Child                 dad could we     1
#> 2   Target_Child                could we turn     1
#> 3   Target_Child                 we turn that     1
#> 4   Target_Child                  out out out     2
#> 5   Target_Child               them them them     3
#> 6   Target_Child                  can you fix     1
#> 7   Target_Child                 you fix them     1
#> 8   Target_Child                 ga sjemm out     1
#> 9   Target_Child            what's over there     1
#> 10  Target_Child                     can I go     3
#> 11  Target_Child                    I go down     1
#> 12  Target_Child                go down there     1
#> 13  Target_Child                  I like that     1
#> 14  Target_Child                first give me     1
#> 15  Target_Child                 give me some     1
#> 16  Target_Child                me some right     1
#> 17  Target_Child             some right there     1
#> 18  Target_Child                   can I have     1
#> 19  Target_Child                  I have some     1
#> 20  Target_Child                 have some of     1
#> 21  Target_Child                 some of that     1
#> 22  Target_Child               Brian in there     1
#> 23  Target_Child                     oh a dog     1
#> 24  Target_Child                    can I get     1
#> 25  Target_Child                   I get down     1
#> 26  Target_Child                where are you     2
#> 27  Target_Child                   ma dawe te     1
#> 28  Target_Child                 what is that     1
#> 29  Target_Child                mami that too     1
#> 30  Target_Child                     oh up up     1
#> 31  Target_Child                   ah wah wah     1
#> 32  Target_Child               where you mama     1
#> 33  Target_Child                  you you you     1
#> 34  Target_Child                  are you Mom     1
#> 35  Target_Child                     go go go     2
#> 36  Target_Child                 mommy get up     1
#> 37  Target_Child                who wanna get     1
#> 38  Target_Child                 wanna get up     1
#> 39  Target_Child                   oh I click     1
#> 40  Target_Child                   do too too     1
#> 41  Target_Child                   I have toy     1
#> 42  Target_Child                have toy them     1
#> 43        Father                wanna give me     1
#> 44        Father                    give me a     5
#> 45        Father                    me a kiss     5
#> 46        Father                 come on give     4
#> 47        Father                   on give me     4
#> 48        Father              give your daddy     1
#> 49        Father                 your daddy a     1
#> 50        Father                 daddy a kiss     1
#> 51        Father                  mm nice fiu     1
#> 52        Father      taperecorder over there     1
#> 53        Father                  do you have     1
#> 54        Father                you have some     1
#> 55        Father               have some nice     1
#> 56        Father             some nice little     1
#> 57        Father           nice little things     1
#> 58        Father             little things to     1
#> 59        Father                things to say     1
#> 60        Father                    to say to     1
#> 61        Father                    say to it     1
#> 62        Father                 if you press     1
#> 63        Father               you press that     1
#> 64        Father             press that it'll     1
#> 65        Father            that it'll really     1
#> 66        Father           it'll really munch     1
#> 67        Father          really munch things     1
#> 68        Father              munch things up     1
#> 69        Father         that will definitely     1
#> 70        Father        will definitely munch     1
#> 71        Father          definitely munch up     1
#> 72        Father                  munch up up     1
#> 73        Father                    now if we     1
#> 74        Father                if we covered     1
#> 75        Father              we covered this     1
#> 76        Father              covered this up     1
#> 77        Father                 this up we'd     1
#> 78        Father                up we'd hafta     1
#> 79        Father              we'd hafta have     1
#> 80        Father                 hafta have a     1
#> 81        Father                  have a hole     1
#> 82        Father                  a hole hole     1
#> 83        Father             still still need     1
#> 84        Father                 still need a     1
#> 85        Father                need a little     1
#> 86        Father                a little hole     1
#> 87        Father              little hole for     1
#> 88        Father                hole for that     1
#> 89        Father               for that built     1
#> 90        Father                that built in     1
#> 91        Father                built in that     1
#> 92        Father                 in that mike     1
#> 93        Father                that mike huh     1
#> 94        Father                other wise we     1
#> 95        Father                wise we could     1
#> 96        Father                  we could be     1
#> 97        Father           could be recording     1
#> 98        Father                     do do do     4
#> 99        Father                 you wanna go     1
#> 100       Father                 wanna go out     2
#> 101       Father                   go out the     1
#> 102       Father               out the window     2
#> 103       Father                 you going to     4
#> 104       Father                 going to see     2
#> 105       Father                     to see a     2
#> 106       Father                  see a mouse     2
#> 107       Father                  a mouse out     1
#> 108       Father               mouse out that     1
#> 109       Father              out that window     1
#> 110       Father                  there is no     1
#> 111       Father                 what is that     1
#> 112       Father                is that Honey     1
#> 113       Father                you know what     4
#> 114       Father               know what that     4
#> 115       Father                 what that is     4
#> 116       Father                okay you know     1
#> 117       Father                  that is the     1
#> 118       Father                  is the case     1
#> 119       Father                 the case for     1
#> 120       Father       cassette tape actually     1
#> 121       Father                     you go a     1
#> 122       Father                    go a cute     1
#> 123       Father                a cute little     1
#> 124       Father            cute little belly     1
#> 125       Father                   now do you     1
#> 126       Father                  do you know     1
#> 127       Father             he's shaking his     2
#> 128       Father             shaking his head     2
#> 129       Father               his head uhhuh     1
#> 130       Father           that's the firemen     3
#> 131       Father             okay now they're     1
#> 132       Father             now they're gone     1
#> 133       Father                  on the step     1
#> 134       Father                you like that     1
#> 135       Father              like that noise     1
#> 136       Father                 his head yes     2
#> 137       Father                  and then he     1
#> 138       Father                then he looks     1
#> 139       Father              he looks around     2
#> 140       Father               you think it's     1
#> 141       Father                 think it's a     1
#> 142       Father                 it's a doggy     1
#> 143       Father                    is that a     1
#> 144       Father                 that a doggy     1
#> 145       Father                a doggy Honey     1
#> 146       Father                   or is he's     1
#> 147       Father                    is he's a     1
#> 148       Father                 he's a kitty     1
#> 149       Father                it's all done     1
#> 150       Father           yeah it's finished     1
#> 151       Father            it's finished now     1
#> 152       Father                 come on back     1
#> 153       Father                 on back over     1
#> 154       Father               back over here     1
#> 155       Father                over here and     1
#> 156       Father               here and let's     1
#> 157       Father               and let's play     1
#> 158       Father              let's play with     1
#> 159       Father               play with your     1
#> 160       Father             with your pillow     1
#> 161       Father                now you wanna     1
#> 162       Father                you wanna try     1
#> 163       Father                 wanna try to     1
#> 164       Father                   try to pop     1
#> 165       Father                  to pop that     1
#> 166       Father                 pop that out     2
#> 167       Father                  that out at     1
#> 168       Father                  out at that     1
#> 169       Father                  at that end     1
#> 170       Father                 or you could     1
#> 171       Father                you could pop     1
#> 172       Father               could pop that     1
#> 173       Father                   and see if     1
#> 174       Father                   see if you     1
#> 175       Father               if you started     1
#> 176       Father          you started getting     1
#> 177       Father         started getting into     1
#> 178       Father              getting into it     1
#> 179       Father                    no if you     1
#> 180       Father                   if you did     1
#> 181       Father                 you did that     1
#> 182       Father                did that that     1
#> 183       Father              that that would     1
#> 184       Father            that would really     1
#> 185       Father            would really mess     1
#> 186       Father               really mess it     1
#> 187       Father                   mess it up     2
#> 188       Father              for example let     1
#> 189       Father               example let me     1
#> 190       Father                   let me see     1
#> 191       Father                    me see if     1
#> 192       Father                  see if that     1
#> 193       Father            if that counter's     1
#> 194       Father         that counter's going     1
#> 195       Father              well it's going     1
#> 196       Father               it's going see     1
#> 197       Father               so don't learn     1
#> 198       Father               don't learn to     1
#> 199       Father                learn to mess     1
#> 200       Father                   to mess it     1
#> 201       Father                    it up yet     1
#> 202       Father                 I wonder how     1
#> 203       Father              wonder how well     1
#> 204       Father                  how well it     1
#> 205       Father                well it would     1
#> 206       Father              it would record     1
#> 207       Father            would record yeah     1
#> 208       Father                   te mit mit     1
#> 209       Father              mit mit csinals     1
#> 210       Father            mit csinals avval     1
#> 211       Father                te esze'l meg     2
#> 212       Father             meg eszlek teged     1
#> 213       Father               okay now those     1
#> 214       Father                now those are     1
#> 215       Father                those are the     2
#> 216       Father                are the knobs     1
#> 217       Father               the knobs that     1
#> 218       Father                 knobs that I     1
#> 219       Father                that I really     1
#> 220       Father               I really don't     1
#> 221       Father            really don't want     1
#> 222       Father               don't want you     1
#> 223       Father                  want you to     1
#> 224       Father                 you to break     1
#> 225       Father            ezeket nem szabad     2
#> 226       Father                  he wants to     1
#> 227       Father                wants to look     1
#> 228       Father                  to look out     1
#> 229       Father                 look out the     1
#> 230       Father             the window again     1
#> 231       Father              this is evening     1
#> 232       Father              is evening time     1
#> 233       Father                 and he's got     1
#> 234       Father                  he's got an     1
#> 235       Father                 got an empty     1
#> 236       Father              an empty bottle     1
#> 237       Father            empty bottle here     1
#> 238       Father                and he thinks     1
#> 239       Father               he thinks it's     1
#> 240       Father             thinks it's very     1
#> 241       Father        it's very interesting     1
#> 242       Father                 and we're up     1
#> 243       Father                  we're up on     1
#> 244       Father                    up on the     1
#> 245       Father                   on the bed     1
#> 246       Father                 the bed with     1
#> 247       Father                 bed with the     1
#> 248       Father        with the taperecorder     1
#> 249       Father                  we just got     1
#> 250       Father                 just got oop     1
#> 251       Father                  yeah well I     1
#> 252       Father                well I really     1
#> 253       Father                I really wish     1
#> 254       Father              really wish you     1
#> 255       Father            wish you wouldn't     1
#> 256       Father            you wouldn't bang     1
#> 257       Father                 that means I     1
#> 258       Father                means I wanna     1
#> 259       Father                   I wanna go     1
#> 260       Father            here's the nipple     1
#> 261       Father               the nipple and     1
#> 262       Father           nipple and there's     1
#> 263       Father              and there's the     1
#> 264       Father           there's the bottle     1
#> 265       Father              they come apart     1
#> 266       Father              come apart like     1
#> 267       Father              apart like that     1
#> 268       Father               would you like     4
#> 269       Father                  you like to     3
#> 270       Father                like to maybe     1
#> 271       Father                 now Ross are     1
#> 272       Father                 Ross are you     1
#> 273       Father                are you going     2
#> 274       Father                going to talk     1
#> 275       Father                   to talk to     1
#> 276       Father                  talk to our     1
#> 277       Father          to our taperecorder     1
#> 278       Father                   on a piece     1
#> 279       Father                   a piece of     1
#> 280       Father                 piece of rye     1
#> 281       Father            what's wrong with     1
#> 282       Father               wrong with him     1
#> 283       Father                okay he waves     1
#> 284       Father                  he waves he     1
#> 285       Father                waves he nods     1
#> 286       Father                  he nods his     1
#> 287       Father                nods his head     1
#> 288       Father                and that's in     1
#> 289       Father          that's in agreement     1
#> 290       Father            in agreement with     1
#> 291       Father            agreement with me     1
#> 292       Father                  with me the     1
#> 293       Father                   me the Dad     1
#> 294       Father                and your nose     1
#> 295       Father                 your nose is     3
#> 296       Father                    nose is a     3
#> 297       Father                    is a Ross     1
#> 298       Father                  a Ross Ross     1
#> 299       Father                    is a mess     2
#> 300       Father                   can I have     1
#> 301       Father                  I have some     1
#> 302       Father               have some yeah     1
#> 303       Father          we're pretending we     1
#> 304       Father          pretending we drink     1
#> 305       Father                 we drink and     1
#> 306       Father               drink and some     1
#> 307       Father                  and some of     1
#> 308       Father                 some of this     1
#> 309       Father                 of this here     1
#> 310       Father                  can I drink     1
#> 311       Father                 I drink some     1
#> 312       Father              drink some more     1
#> 313       Father               some more milk     1
#> 314       Father                you want some     1
#> 315       Father               want some more     1
#> 316       Father                 have a drink     1
#> 317       Father               a drink out_of     1
#> 318       Father            drink out_of here     1
#> 319       Father                like to drink     2
#> 320       Father              to drink out_of     2
#> 321       Father            drink out_of this     1
#> 322       Father              out_of this too     1
#> 323       Father           you're being silly     1
#> 324       Father                you'd like to     1
#> 325       Father               drink out_of a     1
#> 326       Father                 out_of a cup     1
#> 327       Father               a cup wouldn't     1
#> 328       Father             cup wouldn't you     1
#> 329       Father                I think we're     1
#> 330       Father        think we're recording     1
#> 331       Father          we're recording now     1
#> 332       Father              yes Rossy Honey     1
#> 333       Father             Rossy Honey baby     1
#> 334       Father               what would you     2
#> 335       Father               you like Honey     1
#> 336       Father                  you like me     1
#> 337       Father                   like me to     1
#> 338       Father                   me to come     1
#> 339       Father                   to come in     1
#> 340       Father                  come in and     1
#> 341       Father                   in and say     1
#> 342       Father                and say hello     1
#> 343       Father                 say hello to     1
#> 344       Father                 hello to you     1
#> 345       Father                  what do you     1
#> 346       Father                  do you want     1
#> 347       Father                  you want me     1
#> 348       Father                   want me to     1
#> 349       Father                     me to do     1
#> 350       Father                     to do in     1
#> 351       Father                   do in here     1
#> 352       Father                 in here with     1
#> 353       Father                here with you     1
#> 354       Father                 do you wanna     1
#> 355       Father               you wanna come     2
#> 356       Father               wanna come out     1
#> 357       Father                come out here     1
#> 358       Father               does that mean     1
#> 359       Father                that mean you     1
#> 360       Father                 mean you out     1
#> 361       Father                  you out out     1
#> 362       Father                  out out out     1
#> 363       Father                I'll put this     1
#> 364       Father              put this around     1
#> 365       Father              this around you     1
#> 366       Father              around you okay     1
#> 367       Father                  come on Hon     1
#> 368       Father                you take that     1
#> 369       Father                 you can play     1
#> 370       Father                can play with     1
#> 371       Father               play with that     1
#> 372       Father                let's take it     1
#> 373       Father                   take it in     2
#> 374       Father                   it in your     1
#> 375       Father                 in your room     1
#> 376       Father               your room okay     1
#> 377       Father                  we can just     1
#> 378       Father                can just take     1
#> 379       Father                 just take it     1
#> 380       Father                    it in the     1
#> 381       Father                 in the other     1
#> 382       Father               the other room     1
#> 383       Father              other room okay     1
#> 384       Father                you going you     1
#> 385       Father              going you going     1
#> 386       Father              you going going     1
#> 387       Father               going going to     1
#> 388       Father               going to break     1
#> 389       Father                to break that     1
#> 390       Father               break that toy     1
#> 391       Father                 that toy toy     1
#> 392       Father                  toy toy toy     1
#> 393       Father                 going to get     1
#> 394       Father                to get mérges     1
#> 395       Father                get mérges at     1
#> 396       Father                 mérges at it     1
#> 397       Father                 shall we put     1
#> 398       Father                    we put on     2
#> 399       Father                put on diaper     2
#> 400       Father              on diaper Honey     1
#> 401       Father                     oh hi hi     1
#> 402       Father                     hi hi hi     6
#> 403       Father                 what did you     1
#> 404       Father                   did you do     1
#> 405       Father                    you do to     1
#> 406       Father                     do to it     1
#> 407       Father                   to it Ross     1
#> 408       Father                 knock it out     1
#> 409       Father                  is it still     1
#> 410       Father               it still going     1
#> 411       Father                yeah it looks     1
#> 412       Father                it looks like     1
#> 413       Father             looks like we're     1
#> 414       Father             like we're still     1
#> 415       Father            we're still going     1
#> 416       Father             still going here     1
#> 417       Father              let's check the     1
#> 418       Father            check the battery     1
#> 419       Father                what does the     1
#> 420       Father             does the battery     1
#> 421       Father              the battery say     1
#> 422       Father                  like to put     1
#> 423       Father                  to put that     1
#> 424       Father                  put that on     1
#> 425       Father                that on there     1
#> 426       Father                    put it on     2
#> 427       Father                 let's put it     1
#> 428       Father                   it on here     1
#> 429       Father                  come on you     1
#> 430       Father                on you almost     1
#> 431       Father            you almost messed     1
#> 432       Father           almost messed this     1
#> 433       Father               messed this up     1
#> 434       Father     this Tower_of_Hanoi here     1
#> 435       Father                well that's a     1
#> 436       Father                 that's a big     1
#> 437       Father                    a big one     1
#> 438       Father                 pick the big     1
#> 439       Father                  the big one     1
#> 440       Father                   big one up     1
#> 441       Father                     no no no     2
#> 442       Father                 yeah you can     1
#> 443       Father                  you can get     1
#> 444       Father                 can get that     1
#> 445       Father                 get that one     1
#> 446       Father                  that one on     1
#> 447       Father              okay you remove     1
#> 448       Father              you remove that     1
#> 449       Father              remove that one     1
#> 450       Father              means that that     1
#> 451       Father                oh that's his     1
#> 452       Father             that's his first     1
#> 453       Father                his first use     1
#> 454       Father                 first use of     1
#> 455       Father                  use of mami     1
#> 456       Father                  of mami and     1
#> 457       Father                 mami and you     1
#> 458       Father                 and you give     1
#> 459       Father                  you give to     1
#> 460       Father                 give to else     1
#> 461       Father                 to else else     1
#> 462       Father                it means take     1
#> 463       Father                means take it     1
#> 464       Father             that's very nice     1
#> 465       Father              well that's was     1
#> 466       Father               that's was his     1
#> 467       Father                was his first     1
#> 468       Father               his first time     1
#> 469       Father              first time I've     1
#> 470       Father               time I've ever     1
#> 471       Father              I've ever heard     1
#> 472       Father               ever heard him     1
#> 473       Father                heard him use     1
#> 474       Father                   him use it     1
#> 475       Father                  use it that     1
#> 476       Father                  it that way     1
#> 477       Father             okay now where's     1
#> 478       Father              now where's the     1
#> 479       Father              where's the big     1
#> 480       Father                 the big ring     2
#> 481       Father               here's the big     1
#> 482       Father                  how it fits     1
#> 483       Father                 it fits side     1
#> 484       Father               fits side that     1
#> 485       Father                   I put that     1
#> 486       Father                 put that one     1
#> 487       Father               that one there     1
#> 488       Father                you find that     1
#> 489       Father        find that interesting     1
#> 490       Father               I'm going turn     1
#> 491       Father               going turn you     1
#> 492       Father                 turn you off     1
#> 493       Father                  you off for     1
#> 494       Father                    off for a     1
#> 495       Father                 for a second     1
#> 496       Father              a second unless     1
#> 497       Father            second unless you     1
#> 498       Father              unless you have     1
#> 499       Father           you have something     1
#> 500       Father          have something more     1
#> 501       Father            something more to     1
#> 502       Father                  more to say     1
#> 503       Father           means something or     1
#> 504       Father           something or other     1
#> 505       Father                  you want to     4
#> 506       Father                 want to come     4
#> 507       Father                   to come on     1
#> 508       Father                   come on up     2
#> 509       Father                   what can I     1
#> 510       Father                     can I do     1
#> 511       Father                      I do to     1
#> 512       Father                   do to help     1
#> 513       Father                  to help you     1
#> 514       Father                    I you you     1
#> 515       Father                 that's a bad     1
#> 516       Father                   a bad idea     1
#> 517       Father                  bad idea to     1
#> 518       Father                 idea to have     1
#> 519       Father                    to have a     1
#> 520       Father                 have a there     1
#> 521       Father           Ross just finished     1
#> 522       Father         just finished saying     1
#> 523       Father        finished saying don't     1
#> 524       Father           saying don't about     1
#> 525       Father            don't about seven     1
#> 526       Father            about seven times     1
#> 527       Father                he said don't     1
#> 528       Father                and back back     1
#> 529       Father               back back back     3
#> 530       Father                   to him and     1
#> 531       Father                  him and say     1
#> 532       Father                   and say no     1
#> 533       Father              don't like that     1
#> 534       Father                and then he'd     1
#> 535       Father              then he'd smile     1
#> 536       Father               he'd smile and     1
#> 537       Father                 smile and he     1
#> 538       Father                 and he would     1
#> 539       Father             he's standing in     1
#> 540       Father              standing in his     1
#> 541       Father                in his little     1
#> 542       Father        his little Baby_Tenda     1
#> 543       Father              and he's acting     1
#> 544       Father               he's acting up     1
#> 545       Father               acting up like     1
#> 546       Father                    up like a     1
#> 547       Father                  like a pain     1
#> 548       Father               well I suppose     1
#> 549       Father                 I suppose we     1
#> 550       Father              suppose we have     1
#> 551       Father                we have plunk     1
#> 552       Father               have plunk the     1
#> 553       Father              plunk the whole     1
#> 554       Father               the whole baby     1
#> 555       Father             whole baby don't     1
#> 556       Father                baby don't we     1
#> 557       Father                  we might we     1
#> 558       Father               might we might     1
#> 559       Father                   or else we     1
#> 560       Father                else we let's     1
#> 561       Father                we let's just     1
#> 562       Father              let's just wash     1
#> 563       Father                just wash his     1
#> 564       Father               wash his hands     1
#> 565       Father                his hands and     1
#> 566       Father               hands and face     1
#> 567       Father                  at any rate     1
#> 568       Father                  any rate he     1
#> 569       Father               rate he really     1
#> 570       Father                he really got     1
#> 571       Father                 really got a     1
#> 572       Father                    got a big     1
#> 573       Father                   a big kick     1
#> 574       Father              big kick out_of     1
#> 575       Father               kick out_of it     1
#> 576       Father     definitely formal verbal     1
#> 577       Father           formal verbal play     1
#> 578       Father                Mary wants to     1
#> 579       Father                wants to tell     1
#> 580       Father                  to tell the     1
#> 581       Father               tell the story     1
#> 582       Father              the story about     1
#> 583       Father              story about the     1
#> 584       Father            about the diapers     1
#> 585       Father             April May eighth     1
#> 586       Father                May eighth by     1
#> 587       Father                eighth by the     1
#> 588       Father                   by the way     1
#> 589       Father                      I did a     1
#> 590       Father                    did a bad     1
#> 591       Father                     a bad of     1
#> 592       Father             bad of diapering     1
#> 593       Father            of diapering that     1
#> 594       Father          diapering that last     1
#> 595       Father               that last time     1
#> 596       Father                  last time I     1
#> 597       Father                  time I know     1
#> 598       Father                and what'd he     1
#> 599       Father                what'd he say     1
#> 600       Father                  did you poo     1
#> 601       Father                you poo again     1
#> 602       Father                 and that was     1
#> 603       Father                  that was it     1
#> 604       Father               this this this    10
#> 605       Father                  oh now what     1
#> 606       Father                 now what was     1
#> 607       Father                what was that     1
#> 608       Father               was that about     1
#> 609       Father             you had finished     1
#> 610       Father                wanna come on     1
#> 611       Father                you know Ross     1
#> 612       Father            I've got a_lot_of     1
#> 613       Father            got a_lot_of work     1
#> 614       Father             a_lot_of work to     1
#> 615       Father                   work to do     1
#> 616       Father                   I do Honey     1
#> 617       Father      that's the taperecorder     1
#> 618       Father       the taperecorder Honey     1
#> 619       Father                   you see it     1
#> 620       Father                see it moving     1
#> 621       Father                  you see the     2
#> 622       Father               see the needle     2
#> 623       Father            the needle moving     2
#> 624       Father          needle moving there     1
#> 625       Father           moving there Honey     1
#> 626       Father                so what would     1
#> 627       Father                 would you do     1
#> 628       Father                  you do with     1
#> 629       Father                   do with it     1
#> 630       Father                   with it if     1
#> 631       Father                    it if you     1
#> 632       Father                   if you saw     1
#> 633       Father                   you saw it     1
#> 634       Father                saw it moving     1
#> 635       Father               that's for you     1
#> 636       Father                   for you to     1
#> 637       Father               you wanna make     1
#> 638       Father              wanna make that     1
#> 639       Father                   I tell you     1
#> 640       Father                tell you what     1
#> 641       Father                 I'm going to     1
#> 642       Father                 going to put     1
#> 643       Father                   to put you     1
#> 644       Father                 put you over     1
#> 645       Father                you over here     1
#> 646       Father                    hold on a     1
#> 647       Father                  on a second     1
#> 648       Father                you wanna get     2
#> 649       Father                 wanna get up     2
#> 650       Father               where'd you go     1
#> 651       Father                 you go Honey     1
#> 652       Father               did you finish     2
#> 653       Father               you finish did     1
#> 654       Father               finish did you     1
#> 655       Father              you finish your     1
#> 656       Father           finish your bottle     1
#> 657       Father            your bottle again     1
#> 658       Father                   my God you     1
#> 659       Father                 God you just     1
#> 660       Father               you just drank     1
#> 661       Father                just drank it     1
#> 662       Father               drank it right     1
#> 663       Father                it right down     1
#> 664       Father            right down didn't     1
#> 665       Father              down didn't you     1
#> 666       Father             nice clicks Ross     1
#> 667       Father                 you did them     1
#> 668       Father                 did them all     1
#> 669       Father                  them all by     1
#> 670       Father                all by you're     1
#> 671       Father               by you're self     1
#> 672       Father              you're self too     1
#> 673       Father               those are mine     1
#> 674       Father                   to come up     3
#> 675       Father              he's hitting me     1
#> 676       Father                hitting me in     1
#> 677       Father                    me in the     1
#> 678       Father                 in the belly     1
#> 679       Father              trying to climb     1
#> 680       Father                  to climb up     1
#> 681       Father               you can't have     1
#> 682       Father               can't have the     1
#> 683       Father              have the papers     1
#> 684       Father                 you hafta go     1
#> 685       Father                hafta go down     1
#> 686       Father                go down again     1
#> 687       Father               can't you play     1
#> 688       Father                you play with     1
#> 689       Father                play with the     1
#> 690       Father                with the toys     1
#> 691       Father                 the toys you     1
#> 692       Father                toys you have     1
#> 693       Father                I don't think     1
#> 694       Father               don't think so     1
#> 695       Father                think so Ross     1
#> 696       Father                 are the kids     1
#> 697       Father              kids playing at     1
#> 698       Father               playing at the     1
#> 699       Father                at the school     1
#> 700       Father                  Ross look I     1
#> 701       Father                  look I mean     1
#> 702       Father               I mean there's     1
#> 703       Father          mean there's really     1
#> 704       Mother                  let's put a     1
#> 705       Mother                 put a diaper     1
#> 706       Mother                  a diaper on     1
#> 707       Mother               diaper on your     1
#> 708       Mother                 on your face     1
#> 709       Mother               your face okay     1
#> 710       Mother                  turn it off     1
#> 711       Mother            I'm getting tired     1
#> 712       Mother             getting tired of     1
#> 713       Mother                 tired of him     1
#> 714       Mother                  he can have     1
#> 715       Mother                can have this     1
#> 716       Mother           have this macaroni     1
#> 717       Mother            you'll hafta pull     1
#> 718       Mother              hafta pull that     1
#> 719       Mother                pull that out     1
#> 720       Mother       essentially she climbs     1
#> 721       Mother                she climbs up     1
#> 722       Mother                 climbs up on     1
#> 723       Mother                    up on the     1
#> 724       Mother                 on the stove     1
#> 725       Mother              which isn't too     1
#> 726       Mother               isn't too good     1
#> 727       Mother                  too good an     1
#> 728       Mother                 good an idea     1
#> 729       Mother                 as you might     1
#> 730       Mother            you might imagine     1
#> 731       Mother            don't don't don't     1
#> 732       Mother                 shall I wash     1
#> 733       Mother                   I wash his     1
#> 734       Mother               wash his hands     1
#> 735       Mother         diaper diaper diaper     3
#> 736       Mother                yeah I'll let     1
#> 737       Mother                 I'll let you     1
#> 738       Mother                   let you up     1
#> 739       Mother                I was sitting     1
#> 740       Mother             was sitting with     1
#> 741       Mother             sitting with the     1
#> 742       Mother          with the diningroom     1
#> 743       Mother         the diningroom table     1
#> 744       Mother                 Ross came in     1
#> 745       Mother                  came in and     1
#> 746       Mother               in and pointed     1
#> 747       Mother               and pointed to     1
#> 748       Mother               pointed to the     1
#> 749       Mother                 to the where     1
#> 750       Mother              the where where     1
#> 751       Mother            where where Brian     1
#> 752       Mother             where Brian puts     1
#> 753       Mother               Brian puts his     1
#> 754       Mother             puts his diapers     1
#> 755       Mother               his diapers on     1
#> 756       Mother              diapers on with     1
#> 757       Mother                    on with a     1
#> 758       Mother                  with a flap     1
#> 759       Mother                    a flap at     1
#> 760       Mother                  flap at the     1
#> 761       Mother                   at the top     1
#> 762       Mother                 and he would     1
#> 763       Mother                he would kind     1
#> 764       Mother                would kind of     1
#> 765       Mother                kind of shake     1
#> 766       Mother                  of shake at     1
#> 767       Mother                 shake at the     1
#> 768       Mother                  at the flap     1
#> 769       Mother                 the flap and     1
#> 770       Mother             flap and garbled     1
#> 771       Mother        and garbled something     1
#> 772       Mother            anyway but anyway     1
#> 773       Mother                but anyway he     1
#> 774       Mother            anyway he babbled     1
#> 775       Mother               he babbled and     1
#> 776       Mother          babbled and babbled     1
#> 777       Mother                and looked at     1
#> 778       Mother                 looked at me     1
#> 779       Mother                    at me and     1
#> 780       Mother                  me and sort     1
#> 781       Mother                  and sort of     1
#> 782       Mother              sort of wiggled     1
#> 783       Mother               of wiggled his     1
#> 784       Mother           wiggled his diaper     1
#> 785       Mother              his diaper flap     1
#> 786       Mother                   and I said     2
#> 787       Mother                  do you want     1
#> 788       Mother                  you want me     1
#> 789       Mother                   want me to     1
#> 790       Mother                 me to change     1
#> 791       Mother               to change your     1
#> 792       Mother          change your diapers     2
#> 793       Mother                and he nodded     1
#> 794       Mother         he nodded vigorously     1
#> 795       Mother nodded vigorously vigorously     1
#> 796       Mother                     so up up     1
#> 797       Mother                     up up up     1
#> 798       Mother                okay let's go     1
#> 799       Mother              let's go change     1
#> 800       Mother               go change your     1
#> 801       Mother              and he preceded     1
#> 802       Mother               he preceded me     1
#> 803       Mother                 I started to     1
#> 804       Mother              started to move     1
#> 805       Mother             to move preceded     1
#> 806       Mother       move preceded preceded     1
#> 807       Mother   preceded preceded preceded     4
#> 808       Mother             into the bedroom     1
#> 809       Mother                  and he went     1
#> 810       Mother                   he went to     1
#> 811       Mother                  went to his     1
#> 812       Mother                to his closet     1
#> 813       Mother               his closet and     1
#> 814       Mother            closet and looked     1
#> 815       Mother                and looked up     1
#> 816       Mother                 looked up at     1
#> 817       Mother                    up at the     1
#> 818       Mother               at the diapers     1
#> 819       Mother                 and that was     1
#> 820       Mother                that was that     1
#> 821       Mother                   did you hm     1
#> 822       Mother            about about about     3
#> 823       Mother           finding his pillow     1
#> 824       Mother              he followed you     1
#> 825       Mother            followed you down     1
#> 826       Mother                 you down the     2
#> 827       Mother              down the stairs     2
#> 828       Mother                the stairs to     2
#> 829       Mother              stairs to let's     1
#> 830       Mother                 to let's see     1
#> 831       Mother           intended to follow     1
#> 832       Mother                to follow you     1
#> 833       Mother              follow you down     1
#> 834       Mother                 stairs to go     1
#> 835       Mother                  to go watch     1
#> 836       Mother          go watch television     1
#> 837       Mother                  and I think     1
#> 838       Mother                  I think you     1
#> 839       Mother               think you were     1
#> 840       Mother             you were calling     1
#> 841       Mother             were calling him     1
#> 842       Mother                 okay and you     1
#> 843       Mother                  and you got     1
#> 844       Mother                   you got to     1
#> 845       Mother                   got to the     1
#> 846       Mother                  to the edge     1
#> 847       Mother                  the edge of     1
#> 848       Mother                  edge of the     1
#> 849       Mother                of the stairs     1
#> 850       Mother               the stairs and     1
#> 851       Mother                stairs and he     1
#> 852       Mother                and he turned     1
#> 853       Mother             he turned around     1
#> 854       Mother                     I was in     1
#> 855       Mother                   was in the     1
#> 856       Mother            in the diningroom     1
#> 857       Mother      the diningroom watching     1
#> 858       Mother      diningroom watching all     1
#> 859       Mother            watching all this     1
#> 860       Mother                  ran back to     1
#> 861       Mother                  back to his     1
#> 862       Mother               to his bedroom     1
#> 863       Mother               his bedroom no     1
#> 864       Mother              bedroom no back     1
#> 865       Mother                   no back to     1
#> 866       Mother                  back to the     1
#> 867       Mother            to the diningroom     1
#> 868       Mother         the diningroom where     1
#> 869       Mother          diningroom where he     1
#> 870       Mother              where he picked     1
#> 871       Mother                 he picked up     1
#> 872       Mother                picked up his     1
#> 873       Mother                up his bottle     1
#> 874       Mother              his bottle then     1
#> 875       Mother               bottle then he     1
#> 876       Mother                  then he ran     1
#> 877       Mother                    he ran to     1