Get utterances.

Query a table with all the words from the selected transcripts, one word (token) per row. Each row has:

Link to view transcript and play any associated media.

Corpus path to transcript.

Utterance sequence number (starts at 0).

Word sequence number within utterance (starts at 0).

Speaker's ID.

Speaker's role.

Utterance postcodes.

Utterance GEMS.

Utterance.

Start time of utterance in associated media.

End time of utterance in associated media.

getUtterances(
  corpusName = NULL,
  corpora = NULL,
  lang = NULL,
  media = NULL,
  age = NULL,
  gender = NULL,
  designType = NULL,
  activityType = NULL,
  groupType = NULL,
  auth = FALSE
)

Arguments

corpusName
Name of corpus to query. For example, to search within the childes corpus, corpus="childes". Note this is a character string value (not a list). Legal values are 'aphasia', 'asd', 'biling', 'ca', 'childes', 'class', 'dementia', 'fluency', 'homebank', 'phon', 'rhd', 'samtale', 'slabank', and 'tbi'.

corpora
Name of corpus/corporas to query. This is a path starting with the corpus name followed by subfolder names leading to a folder for which all transcripts beneath it will be queried. For example, to query all transripts in the MacWhinney and Brown corpus: corpora = list(list('childes', 'Eng-NA', 'MacWhinney'), list('childes', 'Eng-NA', 'Brown')). Legal values can be found by searching the TalkBank browser: https://sla.talkbank.org/TBB.

lang
Query by language. For example, to get transcripts that contain both English and Spanish: lang=list("eng", "spa"). Legal values: 3-letter language codes based on the ISO 639-3 standard.

media
Query by media type. For example, to get transcripts with an associated video recording: media=list("video"). Legal values: "audio" or "video".

age
Query by participant month age range. For example, to get transcripts with target participants who are either 3-12 or 14-18 months old: age=list(list(from="3", to="12"), list(from="14", to="18")). Legal values: integer month ages.

gender
Query by participant gender. For example, to get transcripts with female target participants: gender=list("female"). Legal values: "female" or "male".

designType
Query by design type. For example, to get transcripts from a longitudinal study: designType=list("long") Legal values are "long" for longitudinal studies, "cross" for cross-sectional studies.

activityType
Query by activity type. For example, to get transcripts where the target particpant is engaged in toy play: activityType=list("toyplay"). See the CHAT manual for legal values.

groupType
Query by group type. For example, to get transcripts where the target particpant is hearing limited: groupType=list("HL"). See the CHAT manual for legal values.

auth
Boolean to determine if user should be prompted to authenticate in order to access protected collections.

Examples

getUtterances(corpusName = 'childes',
              corpora = c('childes',
                                   'Eng-NA',
                                   'MacWhinney',
                                   '010411a'))